backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
distro==1.9.0
fastapi==0.136.1
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-httplib2==0.4.0
googleapis-common-protos==1.75.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.15
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.3.0
langchain-core==1.4.0
langchain-openai==1.2.1
langchain-protocol==0.0.15
langgraph==1.2.0
langgraph-checkpoint==4.1.0
langgraph-prebuilt==1.1.0
langgraph-sdk==0.3.14
langsmith==0.8.4
openai==2.36.0
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
proto-plus==1.28.0
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
pyparsing==3.3.2
python-dotenv==1.2.2
python-multipart==0.0.28
PyYAML==6.0.3
regex==2026.5.9
requests==2.34.2
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.13.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.7.0
uuid_utils==0.15.0
uvicorn==0.47.0
xxhash==3.7.0
zstandard==0.25.0
